MenuToggle moved to separate component
